removed routes, controllers, and model files, as they are no longer called or used. All of their functionality is now in Client.
