+++ server/server.js
@@ -26,10 +26,6 @@ import dotenv from 'dotenv';
 // Importing cors middleware.
 import cors from 'cors';
 // Import router and routes.
-import tracksRouter from './routes/tracksRouter.js';
-import artistsRouter from './routes/artistsRouter.js';
-import albumsRouter from './routes/albumsRouter.js';
-
 
 // Load environment variables from .env.server.
 dotenv.config({ path: '.env.server' });
@@ -51,11 +47,6 @@ const corsOptions = {
     optionSuccessStatus: 200,
 };
 app.use(cors(corsOptions));
-
-// Mounting the router middleware to handle routes starting from the root.
-// app.use('/tracks', tracksRouter);
-app.use('/artists', artistsRouter);
-// app.use('/albums', albumsRouter);
 
 app.listen(PORT, () => {
     console.log(`Server listening on port: ${PORT}...`);
